\name{DICOM Lookup Tables}
\alias{dicom.VRfields}
\alias{dicom.fields}
\title{Lookup Tables for DICOM Header Information}
\description{
}
\usage{data(dicom.VRfields)
data(dicom.fields)
}
\format{
}
\source{
  See references.
}
\references{
  Digital Imaging and Communications in Medicine (DICOM)
  \url{http://medical.nema.org}
  \url{http://en.wikipedia.org/wiki/Digital_Imaging_and_Communications_in_Medicine}
}
\seealso{
  \code{\link{dicom.info}}, \code{\link{dicom.separate}}.
}
\keyword{datasets}
